+++ Source/Scenes/Hexus/States/StateAIDecidePass.cpp
@@ -34,8 +34,8 @@ void StateAIDecidePass::onStateEnter(GameState* gameState)
 	StateBase::onStateEnter(gameState);
 
 	CardData* strongestCardInDeck = gameState->opponentData->getStrongestCard();
-	int strongestAttack = (float)(strongestCardInDeck == nullptr ? 1 : strongestCardInDeck->attack);
-	int passIfDiffAboveOrEqual = (int)(strongestAttack * 2.5f);
+	int strongestAttack = strongestCardInDeck == nullptr ? 1 : int(strongestCardInDeck->attack);
+	int passIfDiffAboveOrEqual = int(strongestAttack * 2.5f);
 	CallFunc* stateTransition = nullptr;
 
 	if (gameState->enemyHand->rowCards.size() == 0)
